requirements.txt
Flask==3.0.3
Flask-Session==0.8.0
africas-talking==1.0.6
python-dotenv==1.0.1
Werkzeug==3.0.3
